Adding Login, Index, and using Bootstrap
Creating a Login.html page that is called if you aren't logged in, and adding Index.html, a traditional ''default" page. I'm using bootstrap to add style to the page. if you navigate to http://localhost:8000/, a call to index.html will be made. if a call to index.html is made without being logged in, a redirect to login.html will occur.